pothole-backend/requirements.txt
numpy
tensorflow-cpu
Pillow
onnxruntime